+++ demo/e2e-bdd/pages/alert.po.ts
@@ -1,0 +1,9 @@
+import { $, ElementFinder } from 'protractor';
+
+import { BasePo } from '../shared/pages/basePage.po';
+
+export class AlertPo extends BasePo {
+  templateUrl = 'alerts';
+  contentTitle: ElementFinder = $('.content-header');
+  pageContent: ElementFinder = $('.content-box');
+}