+++ demo/src/app/docs/demo-section-components/demo-examples-section/examples.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
 
 import { ContentSection } from '../../models/content-section.model';
 import { ComponentExample } from '../../models/components-examples.model';
@@ -12,6 +12,13 @@ export class ExamplesComponent {
   examples: ComponentExample[];
 
   constructor(public section: ContentSection) {
-    this.examples = section.content  as ComponentExample[];
+    this.examples = section.content as ComponentExample[];
+  }
+
+  @HostListener('document:click', ['$event'])
+  preventEmptyHrefNav(event: Event) {
+    if (event && event.target && (event.target as Element).getAttribute('href') === '#') {
+      event.preventDefault();
+    }
   }
 }