+++ src/datepicker/bs-datepicker-input.directive.ts
@@ -1,11 +1,9 @@
-import {
-  ChangeDetectorRef,
-  Directive, ElementRef, forwardRef, Host, OnInit, Renderer2
-} from '@angular/core';
+import { ChangeDetectorRef, Directive, ElementRef, forwardRef, Host, Renderer2 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { BsDatepickerComponent } from './bs-datepicker.component';
+import { BsDatepickerDirective } from './bs-datepicker.component';
 import { formatDate } from '../bs-moment/format';
 import { getLocale } from '../bs-moment/locale/locales.service';
+import { BsLocaleService } from './bs-locale.service';
 
 const BS_DATEPICKER_VALUE_ACCESSOR = {
   provide: NG_VALUE_ACCESSOR,
@@ -27,56 +25,73 @@ export class BsDatepickerInputDirective
   implements ControlValueAccessor {
   private _onChange = Function.prototype;
   private _onTouched = Function.prototype;
+  private _value: Date;
 
-  constructor(@Host() private _picker: BsDatepickerComponent,
+  constructor(@Host() private _picker: BsDatepickerDirective,
+              private _localeService: BsLocaleService,
               private _renderer: Renderer2,
-
               private _elRef: ElementRef,
               private changeDetection: ChangeDetectorRef) {
-    this._picker.bsValueChange.subscribe((v: Date) => this._setInputValue(v));
+    // update input value on datepicker value update
+    this._picker.bsValueChange.subscribe((value: Date) => {
+      this._setInputValue(value);
+      if (this._value !== value) {
+        this._value = value;
+        this._onChange(value);
+        this._onTouched();
+      }
+      this.changeDetection.markForCheck();
+    });
+
+    // update input value on locale change
+    this._localeService.localeChange.subscribe(() => {
+      this._setInputValue(this._value);
+    });
   }
 
-  _setInputValue(v: Date): void {
+  _setInputValue(value: Date): void {
     const initialDate = formatDate(
-      v,
+      value,
       this._picker._config.dateInputFormat,
-      this._picker._config.locale
+      this._localeService.currentLocale
     ) || '';
     this._renderer.setProperty(this._elRef.nativeElement, 'value', initialDate);
-    this._onChange(v);
-    this.changeDetection.markForCheck();
   }
 
   onChange(event: any) {
     this.writeValue(event.target.value);
+    this._onChange(this._value);
     this._onTouched();
   }
 
   writeValue(value: Date | string) {
     if (!value) {
-      this._picker.bsValue = null;
+      this._value = null;
     }
-    const _locale = getLocale(this._picker._config.locale);
+    const _localeKey = this._localeService.currentLocale;
+    const _locale = getLocale(_localeKey);
     if (!_locale) {
       throw new Error(
-        `Locale "${this._picker._config
-          .locale}" is not defined, please add it with "defineLocale(...)"`
+        `Locale "${_localeKey}" is not defined, please add it with "defineLocale(...)"`
       );
     }
     if (typeof value === 'string') {
       const date = new Date(_locale.preparse(value));
-      this._picker.bsValue = isNaN(date.valueOf()) ? null : date;
+      this._value = isNaN(date.valueOf()) ? null : date;
     }
 
     if (value instanceof Date) {
-      this._picker.bsValue = value;
+      this._value = value;
     }
+
+    this._picker.bsValue = this._value;
   }
 
   setDisabledState(isDisabled: boolean): void {
     this._picker.isDisabled = isDisabled;
     if (isDisabled) {
       this._renderer.setAttribute(this._elRef.nativeElement, 'disabled', 'disabled');
+
       return;
     }
     this._renderer.removeAttribute(this._elRef.nativeElement, 'disabled');